Expand null-handling to relates functions and reduce code duplication
